+++ src/lib/ProjectM36/TransactionGraph/Merge.hs
@@ -92,3 +92,17 @@ unionMergeDatabaseContextFunctions prefer funcsA funcsB = case prefer of
   PreferFirst -> pure $ HS.union funcsA funcsB
   PreferSecond -> pure $ HS.union funcsB funcsA
   PreferNeither -> pure $ HS.union funcsA funcsB
+
+unionMergeRegisteredQueries :: MergePreference -> RegisteredQueries -> RegisteredQueries -> Either MergeError RegisteredQueries
+unionMergeRegisteredQueries prefer regQsA regQsB =
+  case prefer of
+    PreferFirst -> pure (M.union regQsA regQsB)
+    PreferSecond -> pure (M.union regQsB regQsA)
+    PreferNeither -> do
+      let isect = M.filter id $ M.mapWithKey (\qname val -> M.lookup qname regQsB /= Just val) (M.intersection regQsA regQsB)
+      --if the values in the intersection are the same, we can merge them      
+      if M.null isect then
+        pure (M.union regQsA regQsB)
+        else
+        Left (StrategyViolatesRegisteredQueryMergeError (M.keys isect))
+